fix: Support attendances over 65k

+++ src/Migrations/2019_01_21_100000_create_events_table.php
@@ -15,7 +15,7 @@ class CreateEventsTable extends Migration
             $t->dateTime('startDateTime');
             $t->dateTime('endDateTime');
             $t->mediumText('description');
-            $t->smallInteger('historicAttendance')->nullable();
+            $t->unsignedInteger('historicAttendance')->nullable();
             $t->dateTime('lastUpdated');
             $t->boolean('isCancelled');
             $t->boolean('isCurrent');
